+++ packages/pigeon/platform_tests/shared_test_plugin_code/test/multiple_arity_test.dart
@@ -14,21 +14,27 @@ import 'multiple_arity_test.mocks.dart';
 void main() {
   test('multiple arity', () async {
     final BinaryMessenger mockMessenger = MockBinaryMessenger();
-    when(mockMessenger.send(
-            'dev.flutter.pigeon.pigeon_integration_tests.MultipleArityHostApi.subtract',
-            any))
-        .thenAnswer((Invocation realInvocation) async {
-      final Object input = MultipleArityHostApi.pigeonChannelCodec
-          .decodeMessage(realInvocation.positionalArguments[1] as ByteData?)!;
+    when(
+      mockMessenger.send(
+        'dev.flutter.pigeon.pigeon_integration_tests.MultipleArityHostApi.subtract',
+        any,
+      ),
+    ).thenAnswer((Invocation realInvocation) async {
+      final Object input =
+          MultipleArityHostApi.pigeonChannelCodec.decodeMessage(
+            realInvocation.positionalArguments[1] as ByteData?,
+          )!;
       final List<Object?> args = input as List<Object?>;
       final int x = (args[0] as int?)!;
       final int y = (args[1] as int?)!;
-      return MultipleArityHostApi.pigeonChannelCodec
-          .encodeMessage(<Object>[x - y]);
+      return MultipleArityHostApi.pigeonChannelCodec.encodeMessage(<Object>[
+        x - y,
+      ]);
     });
 
-    final MultipleArityHostApi api =
-        MultipleArityHostApi(binaryMessenger: mockMessenger);
+    final MultipleArityHostApi api = MultipleArityHostApi(
+      binaryMessenger: mockMessenger,
+    );
     final int result = await api.subtract(30, 10);
     expect(result, 20);
   });