+++ packages/web_benchmarks/testing/test_app/benchmark/test_infra/common.dart
@@ -11,5 +11,5 @@ enum BenchmarkName {
   appScroll,
   appTap,
   simpleBenchmarkPathCheck,
-  simpleCompilationCheck;
+  simpleCompilationCheck,
 }