+++ script/tool/lib/src/common/core.dart
@@ -92,6 +92,7 @@ final Map<Version, Version> _dartSdkForFlutterSdk = <Version, Version>{
   Version(3, 29, 3): Version(3, 7, 2),
   Version(3, 32, 0): Version(3, 8, 0),
   Version(3, 32, 8): Version(3, 8, 1),
+  Version(3, 35, 0): Version(3, 9, 0),
 };
 
 /// Returns the version of the Dart SDK that shipped with the given Flutter